Phylogenetic_based/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
biopython=1.78=py36h8f6f2f9_1
ca-certificates=2020.12.5=ha878542_0
certifi=2020.12.5=py36h5fab9bb_1
dendropy=4.5.2=pyh3252c3a_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=9.3.0=hff62375_18
libgfortran5=9.3.0=hff62375_18
libgomp=9.3.0=h2828fa1_18
liblapack=3.9.0=8_openblas
libopenblas=0.3.12=pthreads_h4812303_1
libstdcxx-ng=9.1.0=hdf63c60_0
ncurses=6.2=he6710b0_1
numpy=1.19.5=py36h2aa4a07_1
openssl=1.1.1k=h7f98852_0
pip=21.0.1=py36h06a4308_0
python=3.6.12=hcff3b4d_2
python_abi=3.6=1_cp36m
readline=8.1=h27cfd23_0
setuptools=52.0.0=py36h06a4308_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
